+++ XFaction/Initialize.lua
@@ -2,6 +2,7 @@ local XFG, G = unpack(select(2, ...))
 local LogCategory = 'Initialize'
 
 function XFG:Init()
+	XFG.WoW = WoWClient:new(); XFG.WoW:Initialize()
 	XFG.Factions = FactionCollection:new(); XFG.Factions:Initialize()
 	XFG.Realms = RealmCollection:new(); XFG.Realms:Initialize()
 	XFG.Guilds = GuildCollection:new(); XFG.Guilds:Initialize()